requirements.txt
Django==5.2.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
psycopg2-binary==2.9.11
google-auth==2.37.0
PyJWT==2.10.1
drf-spectacular==0.29.0
